+++ read.h
@@ -5,5 +5,5 @@
 
 #include "status.h"
 
-bool read_pattern(Status *const status, char *const buf, const int size);
+bool read_pattern(Status *const status, char *const buf, const size_t size);
 #endif /* READ_H */
